+++ database/migrations/2019_05_26_144715_create_works_table.php
@@ -15,18 +15,18 @@ class CreateWorksTable extends Migration
     {
         Schema::create('works', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('studentID')->unsigned();
-            $table->integer('leaderID')->unsigned();
+            $table->bigInteger('studentID')->unsigned();
+            $table->bigInteger('leaderID')->unsigned();
             $table->foreign('studentID')->references('id')->on('users');
             $table->foreign('leaderID')->references('id')->on('users');
             $table->string('themeUkr', 255)->nullable();
             $table->string('themeEn', 255)->nullable();
             $table->string('file', 255)->nullable();
-            $table->integer('dateID')->unsigned()->nullable();
+            $table->bigInteger('dateID')->unsigned()->nullable();
             $table->integer('realPages')->nullable();
             $table->integer('graphicPages')->nullable();
-            $table->integer('rev1')->unsigned()->nullable();
-            $table->integer('rev2')->unsigned()->nullable();
+            $table->bigInteger('rev1')->unsigned()->nullable();
+            $table->bigInteger('rev2')->unsigned()->nullable();
             $table->foreign('dateID')->references('id')->on('dateDef');
             $table->foreign('rev1')->references('id')->on('reviews');
             $table->foreign('rev2')->references('id')->on('reviews');
